test: split llm-core temperature tests
